[FIX] по ревью Kimi: обход валидации через переполнение, SD-восстановление, нулевой стейт регистров
- закрыт обход валидации 16-битным усечением: /water plant65536 давал
  id=0, 65640ml пролезал под cap. Три слоя: cap 6 цифр в парсере,
  checkPlantCommandBounds на ESP (внятное сообщение), long вместо int
  в isValidPlantCommand на Mega
- AwLogging: при ошибке записи хэндл закрывается и переоткрывается —
  вернулось самовосстановление после передёргивания карты
- Valves::init() прогоняет нули в 74HC595: после подачи питания выходы
  случайны до первой защёлки, клапан мог висеть открытым до первого полива
- FlowMeter: инициализаторы полей loopStart/loopNextCheck
- тест на отказ от чисел длиннее 6 цифр; инструкция тестов: прямой
  локальный g++ работает (проблема была в раннере pio test), 22/22 зелёные

+++ src/sketch_auto_watering.ino
@@ -153,7 +153,9 @@ void runDailyCommand() {
 // Валидация команд с plantId/amountMl: id за пределами массива plants — это
 // запись в чужую память и bitWrite с UB (мог открыться случайный клапан),
 // amount без лимита — потоп от опечатки в Telegram.
-bool isValidPlantCommand(int id, int amount) {
+// long, а не int: на AVR int 16-битный, plant65536 усёкся бы до валидного
+// id=0 ещё до проверки.
+bool isValidPlantCommand(long id, long amount) {
   if (id < 0 || id >= PLANTS_AMOUNT) {
     sendTelegram((String)F("Rejected: bad plant id ") + id);
     return false;
@@ -182,8 +184,8 @@ void processEspCommand(JsonDocument& doc) {
   }
 
   if ((String)ESP_COMMAND_WATER_PLANT == command) {
-    int id = doc[F("plantId")];
-    int amount = doc[F("amountMl")];
+    long id = doc[F("plantId")];
+    long amount = doc[F("amountMl")];
     if (!isValidPlantCommand(id, amount)) return;
     logger.buzzerCommand();
     waterPlant(id, amount);
@@ -191,8 +193,8 @@ void processEspCommand(JsonDocument& doc) {
   }
 
   if ((String)ESP_COMMAND_CONFIG_PLANT == command) {
-    int id = doc[F("plantId")];
-    int amount = doc[F("amountMl")];
+    long id = doc[F("plantId")];
+    long amount = doc[F("amountMl")];
     if (!isValidPlantCommand(id, amount)) return;
     global_state.plants[id].dailyAmountMl = amount;
     saveStateEEPROM();
